Seperate ProductAdminController in ProductController
