+++ app/src/main/java/com/codesoom/assignment/controllers/ProductAdminController.java
@@ -1,0 +1,56 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.application.ProductCommandService;
+import com.codesoom.assignment.domain.Product;
+import com.codesoom.assignment.dto.ProductData;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+
+@RestController
+@RequestMapping("/products")
+@CrossOrigin
+public class ProductAdminController {
+    private final ProductCommandService productCommandService;
+
+    public ProductAdminController(ProductCommandService productCommandService) {
+        this.productCommandService = productCommandService;
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    @PreAuthorize("isAuthenticated()")
+    public Product create(
+            @RequestBody @Valid ProductData productData
+    ) {
+        return productCommandService.createProduct(productData);
+    }
+
+    @PatchMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
+    public Product update(
+            @PathVariable Long id,
+            @RequestBody @Valid ProductData productData
+    ) {
+        return productCommandService.updateProduct(id, productData);
+    }
+
+    @DeleteMapping("{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PreAuthorize("isAuthenticated()")
+    public void destroy(
+            @PathVariable Long id
+    ) {
+        productCommandService.deleteProduct(id);
+    }
+}
