+++ app/src/main/java/com/codesoom/assignment/controllers/ProductController.java
@@ -1,15 +1,9 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.application.AuthenticationService;
-import com.codesoom.assignment.application.ProductCommandService;
 import com.codesoom.assignment.application.ProductQueryService;
 import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.dto.ProductData;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -17,15 +11,9 @@ import java.util.List;
 @CrossOrigin
 public class ProductController {
     private final ProductQueryService productQueryService;
-    private final ProductCommandService productCommandService;
 
-    public ProductController(
-            ProductQueryService productQueryService,
-            ProductCommandService productCommandService,
-            AuthenticationService authenticationService
-            ) {
+    public ProductController(ProductQueryService productQueryService) {
         this.productQueryService = productQueryService;
-        this.productCommandService = productCommandService;
     }
 
     @GetMapping
@@ -36,32 +24,5 @@ public class ProductController {
     @GetMapping("{id}")
     public Product detail(@PathVariable Long id) {
         return productQueryService.getProduct(id);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    @PreAuthorize("isAuthenticated()")
-    public Product create(
-            @RequestBody @Valid ProductData productData
-    ) {
-        return productCommandService.createProduct(productData);
-    }
-
-    @PatchMapping("{id}")
-    @PreAuthorize("isAuthenticated()")
-    public Product update(
-            @PathVariable Long id,
-            @RequestBody @Valid ProductData productData
-    ) {
-        return productCommandService.updateProduct(id, productData);
-    }
-
-    @DeleteMapping("{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("isAuthenticated()")
-    public void destroy(
-            @PathVariable Long id
-    ) {
-        productCommandService.deleteProduct(id);
     }
 }
